# perf(fslib): skip symlink logic if the zip file has no symlinks