--- a/packages/yarnpkg-fslib/sources/ZipFS.ts
+++ b/packages/yarnpkg-fslib/sources/ZipFS.ts
@@ -157,6 +157,8 @@ export class ZipFS extends BasePortableFakeFS {
   private readonly listings: Map<PortablePath, Set<Filename>> = new Map();
   private readonly entries: Map<PortablePath, number> = new Map();
 
+  private symlinkCount: number
+
   private readonly fds: Map<number, {cursor: number, p: PortablePath}> = new Map();
   private nextFd: number = 0;
 
@@ -272,6 +274,8 @@ export class ZipFS extends BasePortableFakeFS {
         this.registerListing(p);
       }
     }
+
+    this.symlinkCount = this.libzip.ext.countSymlinks(this.zip);
 
     this.ready = true;
   }
@@ -545,6 +549,14 @@ export class ZipFS extends BasePortableFakeFS {
   }
 
   existsSync(p: PortablePath): boolean {
+    if (!this.ready)
+      throw errors.EBUSY(`archive closed, existsSync '${p}'`);
+
+    if (this.symlinkCount === 0) {
+      const resolvedP = ppath.resolve(PortablePath.root,p);
+      return this.entries.has(resolvedP) || this.listings.has(resolvedP);
+    }
+
     let resolvedP;
 
     try {
@@ -716,7 +728,7 @@ export class ZipFS extends BasePortableFakeFS {
 
     const fileIndex = this.entries.get(resolvedP);
     if (resolveLastComponent && fileIndex !== undefined) {
-      if (this.isSymbolicLink(fileIndex)) {
+      if (this.symlinkCount !== 0 && this.isSymbolicLink(fileIndex)) {
         const target = this.getFileSource(fileIndex).toString() as PortablePath;
         return this.resolveFilename(reason, ppath.resolve(ppath.dirname(resolvedP), target), true);
       } else {
@@ -736,7 +748,7 @@ export class ZipFS extends BasePortableFakeFS {
         throw errors.ENOTDIR(reason);
 
       resolvedP = ppath.resolve(parentP, ppath.basename(resolvedP));
-      if (!resolveLastComponent)
+      if (!resolveLastComponent || this.symlinkCount === 0)
         break;
 
       const index = this.libzip.name.locate(this.zip, resolvedP.slice(1));
@@ -823,6 +835,9 @@ export class ZipFS extends BasePortableFakeFS {
   }
 
   private isSymbolicLink(index: number) {
+    if (this.symlinkCount === 0)
+      return false;
+
     const attrs = this.libzip.file.getExternalAttributes(this.zip, index, 0, 0, this.libzip.uint08S, this.libzip.uint32S);
     if (attrs === -1)
       throw new Error(this.libzip.error.strerror(this.libzip.getError(this.zip)));
@@ -1099,9 +1114,10 @@ export class ZipFS extends BasePortableFakeFS {
     this.registerEntry(resolvedP, index);
 
     const rc = this.libzip.file.setExternalAttributes(this.zip, index, 0, 0, this.libzip.ZIP_OPSYS_UNIX, (0o120000 | 0o777) << 16);
-    if (rc === -1) {
+    if (rc === -1)
       throw new Error(this.libzip.error.strerror(this.libzip.getError(this.zip)));
-    }
+
+    this.symlinkCount += 1;
   }
 
   readFilePromise(p: FSPath<PortablePath>, encoding: 'utf8'): Promise<string>;