… .pnp.js
… packages/yarnpkg-libzip/sources/libzipAsync.js
… packages/yarnpkg-libzip/sources/libzipSync.js
… packages/yarnpkg-pnp/sources/hook.js