**What's the problem this PR addresses?**

In order to support symlinks ZipFS has to do extra resolving logic on each part of a filepath, even though it's most likely that the zip file doesn't contain any symlinks, meaning the time spent resolving is wasted.

**How did you fix it?**

Added `zip_ext_count_symlinks` to libzip (https://github.com/arcanis/libzip/pull/1) and used this function to know if the zip file contains symlinks, if it doesn't contain any, skip the symlink resolving.

**Result**
Measured using `perf stat -r 10 -- yarn test:lint`
|before|after|
|--|--|
| 33,524599848 seconds ( +- 0,15% ) | 32,947723121 seconds ( +- 0,13% )

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I have verified that all automated PR checks pass.